% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigArray-package.R
\docType{package}
\name{BigArray-package}
\alias{BigArray-package}
\alias{_PACKAGE}
\title{BigArray: Using bigmemory as a DelayedArray Backend}
\description{
Implements a DelayedArray backend for reading and writing dense matrices in the bigmemory format. Supports bigmemory-backed SummarizedExperiment objects.
}
\keyword{internal}
