% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigRealizationSink-class.R
\name{dim,BigRealizationSink-method}
\alias{dim,BigRealizationSink-method}
\title{Internal functions}
\usage{
\S4method{dim}{BigRealizationSink}(x)
}
\arguments{
\item{x}{BigArraySeed, BigArray, BigRealizationSink}
}
\description{
Internal functions that should not be used called directly.
}
