% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSummarizedExperiment.R
\name{saveBigArraySummarizedExperiment}
\alias{saveBigArraySummarizedExperiment}
\alias{loadBigArraySummarizedExperiment}
\title{Save/Load a BigArray-based SummarizedExperiment object}
\usage{
saveBigArraySummarizedExperiment(
  x,
  dir = tempfile(pattern = "dir"),
  prefix = "",
  replace = FALSE
)

loadBigArraySummarizedExperiment(dir, prefix = "")
}
\arguments{
\item{x}{SummarizedExperiment: All assays will be coerced to BigArray objects.}

\item{dir}{character: Path to directory where to save the data.}

\item{prefix}{character: Optional prefix to add to the saved files. Allows saving more than one object in the same directory}

\item{replace}{logical: Whether to overwrite/delete all files if directory already exists.}
}
\value{
A SummarizedExperiment object where each assay is stored as a BigArray.
}
\description{
Saves a SummarizedExperiment-object do disk where each assay is stored as a BigArray.
Modelled after \link[HDF5Array]{saveHDF5SummarizedExperiment} and \link[HDF5Array]{loadHDF5SummarizedExperiment}
}
\examples{
#### Examples ####

library(SummarizedExperiment)

nrow <- 200
ncol <- 6
counts <- matrix(as.integer(runif(nrow * ncol, 1, 1e4)), nrow)
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
se0 <- SummarizedExperiment(assays=list(counts=counts), colData=colData)
se0

## Save 'se0' as an HDF5-based SummarizedExperiment object:
dir <- tempfile("h5_se0_")
h5_se0 <- saveBigArraySummarizedExperiment(se0, dir)
list.files(dir)

h5_se0
assay(h5_se0)

h5_se0b <- loadBigArraySummarizedExperiment(dir)
h5_se0b
assay(h5_se0b)

#### Sanity checks ####

a1 <- as(as.matrix(iris[,-5]), "BigArray")
a2 <- a1 * 10
a3 <- round(a2)
se1 <- SummarizedExperiment(assays=SimpleList(a1=a1, a2=a2),
                            rowData = iris[,5, drop=FALSE])
se2 <- SummarizedExperiment(assays=SimpleList(a3 = a3),
                            rowData = iris[,5, drop=FALSE])

# Save to a directory
input_dir <- tempfile()
input_se1 <- saveBigArraySummarizedExperiment(se1, dir = input_dir)

# Move and reload
output_dir <- paste0(input_dir, "_moved")
file.rename(input_dir, to = output_dir)
output_se1 <- loadBigArraySummarizedExperiment(output_dir)

# Check reload works
stopifnot(all(a1 ==  assay(output_se1, "a1")),
          all(a2 ==  assay(output_se1, "a2")))

# Check prefixes work
prefix_dir <- tempfile()
A1 <- saveBigArraySummarizedExperiment(se1, dir = prefix_dir, prefix = "A_")
B1 <- saveBigArraySummarizedExperiment(se2, dir = prefix_dir, prefix = "B_", replace = TRUE)

A2 <- loadBigArraySummarizedExperiment(prefix_dir, prefix = "A_")
B2 <- loadBigArraySummarizedExperiment(prefix_dir, prefix = "B_")

stopifnot(all(assay(A1) == assay(A2)),
          all(assay(B1) == assay(B2)))
}
