% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed-aware.R
\name{seed-aware}
\alias{seed-aware}
\alias{colSums2,BigArraySeed-method}
\alias{colProds,BigArraySeed-method}
\alias{colMeans2,BigArraySeed-method}
\alias{colVars,BigArraySeed-method}
\alias{colSds,BigArraySeed-method}
\alias{colRanges,BigArraySeed-method}
\alias{colMins,BigArraySeed-method}
\alias{colMins,BigArray-method}
\alias{colMaxs,BigArraySeed-method}
\alias{colMaxs,BigArray-method}
\title{Matrix summary stats for BigArray}
\usage{
\S4method{colSums2}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colProds}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colMeans2}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colVars}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colSds}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colRanges}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colMins}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colMins}{BigArray}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colMaxs}{BigArraySeed}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)

\S4method{colMaxs}{BigArray}(x, rows = NULL, cols = NULL, na.rm = FALSE, useNames = NA)
}
\arguments{
\item{x}{BigArraySeed.}

\item{rows}{Not supported for BigArray.}

\item{cols}{integer, character or logical: Subset of columns to use.}

\item{na.rm}{logical: Whether do disregards missing values.}

\item{useNames}{Not supported for BigArray.}
}
\value{
A vector of column summary statistics (for colRanges a matrix).
}
\description{
"Seed-aware" matrix summary statistics for BigArray. These should normally not
be called directly, but rather used via the DelayedMatrixStats package: See
examples below.
}
\note{
Requires the biganalytics package.
}
\examples{
library(DelayedMatrixStats)

# Parallel options
DelayedArray:::set_verbose_block_processing(TRUE)
setAutoBlockSize(1e8) # Silly small block size to see what is going on

# Write a large matrix to a BigArray
m <- replicate(n=10, rnorm(10000))
ba <- writeBigArray(m)

# Seed aware method
colSums2(ba)

# Force DelayedMatrixStats to use block processing
colSums2(ba, force_block_processing = TRUE)

# The seed-aware method is only used when the input is a BigArray. E.g. first
# subsetting a BigArray will demote it to a DelayedArray, which can only use
# block-processing
ba_subset <- ba[1:100, 1:5]
class(ba_subset) # No longer a BigArray
colSums2(ba_subset) # Block-processed

# All seed aware
colSums2(ba)
colMins(ba)
colMaxs(ba)
colProds(ba)
colMeans2(ba)
colVars(ba)
colSds(ba)
colRanges(ba)
}
